# tests/contract/api-gateway-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent PPM Platform",
    "description": "AI-native Project Portfolio Management platform with 25 specialized agents",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/agents/{agent_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Get Agent Info",
        "description": "Get information about a specific agent.",
        "operationId": "get_agent_info_v1_agents__agent_id__get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Agent Info V1 Agents  Agent Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}/agents/enabled": {
      "get": {
        "tags": [
          "agent-config"
        ],
        "summary": "Get Enabled Agents For Project",
        "description": "Get list of enabled agents for a specific project.",
        "operationId": "get_enabled_agents_for_project_v1_projects__project_id__agents_enabled_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Enabled Agents For Project V1 Projects  Project Id  Agents Enabled Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/agents/{agent_id}/enabled": {
      "get": {
        "tags": [
          "agent-config"
        ],
        "summary": "Check Agent Enabled For Project",
        "description": "Check if a specific agent is enabled for a project.",
        "operationId": "check_agent_enabled_for_project_v1_projects__project_id__agents__agent_id__enabled_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Check Agent Enabled For Project V1 Projects  Project Id  Agents  Agent Id  Enabled Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}/risks/research": {
      "post": {
        "tags": [
          "risk-research"
        ],
        "summary": "Research Project Risks",
        "description": "Trigger external risk research using the Risk & Issue Management agent.",
        "operationId": "research_project_risks_v1_projects__project_id__risks_research_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskResearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}/compliance/research": {
      "post": {
        "tags": [
          "compliance-research"
        ],
        "summary": "Research Compliance",
        "description": "Trigger external regulatory monitoring using the Compliance agent.",
        "operationId": "research_compliance_v1_projects__project_id__compliance_research_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceResearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplianceResearchRequest": {
        "properties": {
          "domain": {
            "type": "string",
            "minLength": 2,
            "title": "Domain",
            "description": "Project domain or industry."
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region",
            "description": "Region for regulatory monitoring."
          }
        },
        "type": "object",
        "required": [
          "domain"
        ],
        "title": "ComplianceResearchRequest"
      },
      "ComplianceResearchResponse": {
        "properties": {
          "changes_detected": {
            "type": "integer",
            "title": "Changes Detected"
          },
          "regulations_impacted": {
            "type": "integer",
            "title": "Regulations Impacted"
          },
          "impacted_regulations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Impacted Regulations"
          },
          "external_monitoring": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Monitoring"
          },
          "last_check": {
            "type": "string",
            "title": "Last Check"
          }
        },
        "type": "object",
        "required": [
          "changes_detected",
          "regulations_impacted",
          "impacted_regulations",
          "last_check"
        ],
        "title": "ComplianceResearchResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskResearchRequest": {
        "properties": {
          "context": {
            "type": "string",
            "minLength": 3,
            "title": "Context",
            "description": "Project domain or context."
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region",
            "description": "Geographic region for the project."
          },
          "categories": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categories",
            "description": "High-level risk categories."
          }
        },
        "type": "object",
        "required": [
          "context"
        ],
        "title": "RiskResearchRequest"
      },
      "RiskResearchResponse": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "external_risks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "External Risks"
          },
          "added_risks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Added Risks"
          },
          "used_external_research": {
            "type": "boolean",
            "title": "Used External Research"
          },
          "notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "external_risks",
          "added_risks",
          "used_external_research"
        ],
        "title": "RiskResearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}